# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Employee Onboarding HRMS API",
    "version": "1.0.0",
    "description": "Swagger/OpenAPI documentation for the Employee Onboarding HRMS backend. Most endpoints return `{ status, message?, data? }` and protected routes require a JWT bearer token."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local backend"
    },
    {
      "url": "https://bitbyte-backend.onrender.com",
      "description": "Deployed backend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get admin dashboard stats",
        "description": "Roles: admin.",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/v1/super-admin/dashboard": {
      "get": {
        "tags": ["Super Admin"],
        "summary": "Get super admin dashboard stats",
        "description": "Roles: super_admin.",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "description": "Roles: super_admin.",
        "parameters": [
          { "$ref": "#/components/parameters/RoleQuery" },
          { "$ref": "#/components/parameters/StatusQuery" },
          { "$ref": "#/components/parameters/PageQuery" },
          { "$ref": "#/components/parameters/LimitQuery" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/UserListSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List current user notifications",
        "responses": {
          "200": { "$ref": "#/components/responses/NotificationsSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "PageQuery": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 20 }
      },
...
      "StatusQuery": {
        "name": "status",
        "in": "query",
        "schema": { "type": "string" }
      },
      "RoleQuery": {
        "name": "role",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["employee", "admin", "super_admin", "intern"]
        }
      },
...
    },
    "responses": {
      "GenericSuccess": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" }
          }
        }
      },
...
      "UserListSuccess": {
        "description": "Users list response",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/User" }
                        },
                        "pagination": { "$ref": "#/components/schemas/Pagination" }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotificationsSuccess": {
        "description": "Notifications response",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "notifications": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Notification" }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Role or first-login access denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "fail" },
          "message": { "type": "string" },
          "code": { "type": "string" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 100 },
          "pages": { "type": "integer", "example": 5 }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["employee", "admin", "super_admin", "intern"]
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "suspended"]
          },
          "firstLogin": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "recipientId": { "type": "string" },
          "type": { "type": "string" },
          "channel": { "type": "string", "example": "in_app" },
          "subject": { "type": "string" },
          "body": { "type": "string" },
          "status": { "type": "string", "example": "pending" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
...
    }
  }
}